requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.2
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.2
bleach==6.2.0
boto3==1.28.57
botocore==1.31.85
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==5.2.7
django-allauth==0.57.0
django-ckeditor==6.7.0
django-elasticsearch-dsl==8.0
django-js-asset==3.1.2
django-redis==6.0.0
django-storages==1.14.2
django-taggit==6.1.0
django-tailwind==4.2.0
dnspython==2.8.0
elastic-transport==8.17.1
elasticsearch==8.13.0
elasticsearch-dsl==8.18.0
exceptiongroup==1.3.0
fastapi==0.103.1
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
kombu==5.5.4
msgpack==1.1.1
mysqlclient==2.2.0
oauthlib==3.3.1
packaging==25.0
Pillow==10.0.1
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.5.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.1
python3-openid==3.2.0
redis==5.0.1
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.7.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.27.0
text-unidecode==1.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3<2.1,>=1.25.4
uvicorn==0.23.2
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.14
webencodings==0.5.1
zope.interface==8.0.1
